Move sliding window and fix for golden configs
